tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-slack-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: masonry-slack parsing</title>
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="masonry-slack supports only the grammar 'normal | <length-percentage>'.">
<meta name="assert" content="masonry-slack rejects negative <length-percentage>.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("masonry-slack", "auto");

test_invalid_value("masonry-slack", "10");
test_invalid_value("masonry-slack", "10px 20px");
test_invalid_value("masonry-slack", "-1px");
test_invalid_value("masonry-slack", "-10%");
</script>
</body>
</html>
